настроил вывод всплывающего окна добавил новые ресурсные файлы перевел все поинты с флоат на дабл так как все функции которые работают с поинтами используют дабл

+++ src/com/example/MapApp/Main/MyPosition.java
@@ -5,16 +5,16 @@ package com.example.MapApp.Main;
  */
 public class MyPosition implements Point{
 
-    public static float latitude;
-    public static float longitude;
+    public static double latitude;
+    public static double longitude;
 
     @Override
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
     @Override
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 }
