+++ src/com/example/MapApp/Main/Point.java
@@ -4,6 +4,6 @@ package com.example.MapApp.Main;
  * Created by respect on 3/14/14.
  */
 public interface Point {
-    public float getLatitude();
-    public float getLongitude();
+    public double getLatitude();
+    public double getLongitude();
 }
